fix: mark linux exclusive sources with build tags
gopacket/gopacket/afpacket works with Linux only, hence marking
all sources that depend on it improves DX for other platforms such
as macOS.
